requirements.txt
streamlit==1.24.1
umap-learn==0.5.3
transformers==4.30.2
torch==2.0.1
plotly==5.15.0
rouge-score==0.1.2
wordcloud==1.9.2